Implementation de reservation de creneau

+++ sportify/src/main/java/fr/univ/lyon1/m1if/m1if10Grp13/dao/DAOTerrain.java
@@ -75,7 +75,7 @@ public class DAOTerrain implements DAOCrud {
         Long terrainId = null;
 
         // verifier si l'objet passé est de type Long
-        if (object instanceof Long) {
+        if (object instanceof Integer) {
             terrainId = (Long) object;
         }
